BREAKING CHANGE(ClickableArea)
Made ClickableArea inherit HoverableArea, and take the previous role of MouseHelper, which is now called ClickableWindow and inherits ClickableArea, instead of the other way around.

Cleaned up ClickableArea's logic. Fixed (and left the option in as well) autotriggering when entering the actionbox with the button down. Added HoldOut button option.